@@ -13,12 +13,11 @@ part 'local_contacts_state.dart';
 
 class LocalContactsBloc extends Bloc<LocalContactsEvent, LocalContactsState> {
   final LocalContactsRepository localContactsRepository;
-  StreamSubscription _localContactsSubscription;
+  StreamSubscription? _localContactsSubscription;
 
   LocalContactsBloc({
-    @required this.localContactsRepository,
-  })  : assert(localContactsRepository != null),
-        super(LocalContactsInitial());
+    required this.localContactsRepository,
+  }) : super(LocalContactsInitial());
 
   @override
   Stream<LocalContactsState> mapEventToState(LocalContactsEvent event) async* {
@@ -37,7 +36,8 @@ class LocalContactsBloc extends Bloc<LocalContactsEvent, LocalContactsState> {
     return super.close();
   }
 
-  Stream<LocalContactsState> _mapLocalContactsInitialLoadedToState(LocalContactsInitialLoaded event) async* {
+  Stream<LocalContactsState> _mapLocalContactsInitialLoadedToState(
+      LocalContactsInitialLoaded event) async* {
     yield LocalContactsInitial();
     _localContactsSubscription?.cancel();
     _localContactsSubscription = localContactsRepository.contacts().listen(
@@ -51,7 +51,8 @@ class LocalContactsBloc extends Bloc<LocalContactsEvent, LocalContactsState> {
     }
   }
 
-  Stream<LocalContactsState> _mapLocalContactsRefreshedToState(LocalContactsRefreshed event) async* {
+  Stream<LocalContactsState> _mapLocalContactsRefreshedToState(
+      LocalContactsRefreshed event) async* {
     try {
       await localContactsRepository.load();
     } catch (error) {
@@ -59,7 +60,8 @@ class LocalContactsBloc extends Bloc<LocalContactsEvent, LocalContactsState> {
     }
   }
 
-  Stream<LocalContactsState> _mapLocalContactsUpdatedToState(LocalContactsUpdated event) async* {
+  Stream<LocalContactsState> _mapLocalContactsUpdatedToState(
+      LocalContactsUpdated event) async* {
     yield LocalContactsLoadSuccess(contacts: event.contacts);
   }
 }
