@@ -8,7 +8,7 @@ class SimpleBlocObserver extends BlocObserver {
   }
 
   @override
-  void onEvent(Bloc bloc, Object event) {
+  void onEvent(Bloc bloc, Object? event) {
     super.onEvent(bloc, event);
     print('onEvent $bloc: $event');
   }
