@@ -13,7 +13,7 @@ import 'package:webtrit_phone/app.dart';
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(App());
+    await tester.pumpWidget(App(webRegistrationInitialUrl: '', isRegistered: false));
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
